More extensive GPG verification for DNF
Based on known GPG fingerprint value, download the docker GPG key,
verify the fingerprint matches, drop that onto the filesystem, then
point the repo to that file.

This is a round-about way to recreate the functionality that apt_key is
doing. Unfortunately rpm_key doesn't allow adding a key via fingerprint
like the apt_key module.